requirements.txt
fastapi==0.111.0
google-generativeai==0.7.1
loguru==0.7.2
google-cloud-firestore==2.16.1